+++ default.nix
@@ -18,6 +18,8 @@ let
     rhyolite-backend = ./backend;
     rhyolite-backend-db = ./backend-db;
     notify-listen = ./notify-listen;
+    postgresql-simple-class = ./postgresql-simple-class;
+    groundhog-legacy = ./groundhog-legacy;
     rhyolite-common = ./common;
     rhyolite-datastructures = ./datastructures;
     rhyolite-frontend = ./frontend;